moved ISimulationControl to de.uka.ipd.sdq.simulation.core
