+++ bundles/de.uka.ipd.sdq.simucomframework.simulationdock/src/de/uka/ipd/sdq/simucomframework/simulationdock/DispatchingSimulationObserver.java
@@ -2,7 +2,7 @@ package de.uka.ipd.sdq.simucomframework.simulationdock;
 
 import java.util.ArrayList;
 
-import de.uka.ipd.sdq.simulation.IStatusObserver;
+import de.uka.ipd.sdq.simulation.core.IStatusObserver;
 
 public class DispatchingSimulationObserver implements IStatusObserver {
