+++ bundles/de.uka.ipd.sdq.simulation.core/META-INF/MANIFEST.MF
@@ -8,7 +8,8 @@ Require-Bundle: org.apache.log4j;bundle-version="1.2.19",
  org.palladiosimulator.analyzer.workflow.core;bundle-version="5.2.1",
  org.palladiosimulator.recorderframework;bundle-version="5.2.1",
  de.uka.ipd.sdq.probfunction.math;bundle-version="5.2.1",
- de.uka.ipd.sdq.simulation.abstractsimengine;bundle-version="5.2.1"
+ de.uka.ipd.sdq.simulation.abstractsimengine;bundle-version="5.2.1",
+ de.uka.ipd.sdq.simucomframework.simucomstatus;bundle-version="5.2.1"
 Bundle-Vendor: edu.kit.ipd.sdq
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-17
