+++ bundles/de.uka.ipd.sdq.simulation.core/src/de/uka/ipd/sdq/simulation/core/IStatusObserver.java
@@ -1,4 +1,4 @@
-package de.uka.ipd.sdq.simulation;
+package de.uka.ipd.sdq.simulation.core;
 
 /**
  * Interface for observers interested in the progress of the simulation
